+++ Dyalect/Compiler/OpSizeHelper.cs
@@ -91,8 +91,7 @@
             1, //TypeST
             1, //CtorCheck
             0, //Unbox
-            1, //PopAuto
-            0, //CloseAuto
+            0, //Rethrow
         };
     }
 }